Extract device_tab widget into package submodules
- Split 3092-line device_tab.py into focused modules
- Separate capture flows, grid layout, mapping display, and helpers
- Update test imports for new package structure
