+++ keymasq/gui/widgets/device_tab/__init__.py
@@ -1,0 +1,36 @@
+import gi
+
+gi.require_version("Gtk", "4.0")
+
+from gi.repository import GLib  # pyright: ignore[reportAttributeAccessIssue]
+
+from keymasq.gui.widgets.device_tab.capture_helpers import (
+    _make_capture_status_row,
+    _set_capture_status,
+)
+from keymasq.gui.widgets.device_tab.grid import (
+    _grouped_analog_inputs,
+    _ordered_analog_inputs,
+)
+from keymasq.gui.widgets.device_tab.mapping_display import (
+    _char_middle_shorten_text,
+    _compact_exec_summary,
+    _display_action_summary,
+)
+from keymasq.gui.widgets.device_tab.tab import (
+    DeviceTab,
+    session_request_async,
+)
+
+__all__ = [
+    "DeviceTab",
+    "GLib",
+    "_char_middle_shorten_text",
+    "_compact_exec_summary",
+    "_display_action_summary",
+    "_grouped_analog_inputs",
+    "_make_capture_status_row",
+    "_ordered_analog_inputs",
+    "_set_capture_status",
+    "session_request_async",
+]
